TestHarness adjustments for #402
