+++ tests/test_score_flux_yn/test_score_flux_yn.py
@@ -2,9 +2,9 @@
 
 import sys
 sys.path.insert(0, '..')
-from testing_harness import TestHarness
+from testing_harness import HashedTestHarness
 
 
 if __name__ == '__main__':
-    harness = TestHarness('statepoint.10.*', True)
+    harness = HashedTestHarness('statepoint.10.*', True)
     harness.execute_test()
