+++ tests/test_statepoint_sourcesep/test_statepoint_sourcesep.py
@@ -18,5 +18,5 @@ class SourcepointTestHarness(TestHarness):
 
 
 if __name__ == '__main__':
-    harness = TestHarness('statepoint.10.*')
+    harness = SourcepointTestHarness('statepoint.10.*')
     harness.execute_test()
